+++ lua/lazyvim/health.lua
@@ -17,7 +17,7 @@ function M.check()
     error("Neovim >= 0.9.0 is required")
   end
 
-  for _, cmd in ipairs({ "git", "rg", { "fd", "fdfind" }, "lazygit" }) do
+  for _, cmd in ipairs({ "git", "rg", { "fd", "fdfind" }, "lazygit", "fzf", "curl" }) do
     local name = type(cmd) == "string" and cmd or vim.inspect(cmd)
     local commands = type(cmd) == "string" and { cmd } or cmd
     ---@cast commands string[]