+++ lua/lazyvim/plugins/colorscheme.lua
@@ -19,6 +19,7 @@ return {
         cmp = true,
         dashboard = true,
         flash = true,
+        fzf = true,
         grug_far = true,
         gitsigns = true,
         headlines = true,