+++ lua/lazyvim/plugins/extras/ai/copilot.lua
@@ -58,67 +58,61 @@ return {
     end,
   },
 
-  -- copilot cmp source
-  {
-    "nvim-cmp",
-    optional = true,
-    dependencies = { -- this will only be evaluated if nvim-cmp is enabled
-      {
-        "zbirenbaum/copilot-cmp",
-        enabled = vim.g.ai_cmp, -- only enable if wanted
-        opts = {},
-        config = function(_, opts)
-          local copilot_cmp = require("copilot_cmp")
-          copilot_cmp.setup(opts)
-          -- attach cmp source whenever copilot attaches
-          -- fixes lazy-loading issues with the copilot cmp source
-          LazyVim.lsp.on_attach(function()
-            copilot_cmp._on_insert_enter({})
-          end, "copilot")
-        end,
-        specs = {
-          {
-            "nvim-cmp",
-            optional = true,
-            ---@param opts cmp.ConfigSchema
-            opts = function(_, opts)
-              table.insert(opts.sources, 1, {
-                name = "copilot",
-                group_index = 1,
-                priority = 100,
-              })
-            end,
-          },
-        },
-      },
-    },
-  },
-
-  -- blink.cmp
-  {
-    "saghen/blink.cmp",
-    optional = true,
-    dependencies = {
-      {
-        "giuxtaposition/blink-cmp-copilot",
-        enabled = vim.g.ai_cmp, -- only enable if needed
-        specs = {
-          {
-            "blink.cmp",
-            optional = true,
-            opts = {
-              sources = {
-                providers = {
-                  copilot = { name = "copilot", module = "blink-cmp-copilot" },
-                },
-                completion = {
-                  enabled_providers = { "copilot" },
+  vim.g.ai_cmp
+      and {
+        -- copilot cmp source
+        {
+          "hrsh7th/nvim-cmp",
+          optional = true,
+          dependencies = { -- this will only be evaluated if nvim-cmp is enabled
+            {
+              "zbirenbaum/copilot-cmp",
+              opts = {},
+              config = function(_, opts)
+                local copilot_cmp = require("copilot_cmp")
+                copilot_cmp.setup(opts)
+                -- attach cmp source whenever copilot attaches
+                -- fixes lazy-loading issues with the copilot cmp source
+                LazyVim.lsp.on_attach(function()
+                  copilot_cmp._on_insert_enter({})
+                end, "copilot")
+              end,
+              specs = {
+                {
+                  "hrsh7th/nvim-cmp",
+                  optional = true,
+                  ---@param opts cmp.ConfigSchema
+                  opts = function(_, opts)
+                    table.insert(opts.sources, 1, {
+                      name = "copilot",
+                      group_index = 1,
+                      priority = 100,
+                    })
+                  end,
                 },
               },
             },
           },
         },
-      },
-    },
-  },
+        {
+          "saghen/blink.cmp",
+          optional = true,
+          dependencies = { "giuxtaposition/blink-cmp-copilot" },
+          opts = {
+            sources = {
+              default = { "copilot" },
+              providers = {
+                copilot = {
+                  name = "copilot",
+                  module = "blink-cmp-copilot",
+                  kind = "Copilot",
+                  score_offset = 100,
+                  async = true,
+                },
+              },
+            },
+          },
+        },
+      }
+    or nil,
 }