+++ lua/lazyvim/plugins/extras/ai/supermaven.lua
@@ -6,6 +6,7 @@ return {
         accept_suggestion = nil, -- handled by nvim-cmp / blink.cmp
       },
       disable_inline_completion = vim.g.ai_cmp,
+      ignore_filetypes = { "bigfile", "snacks_input", "snacks_notif" },
     },
   },
 
@@ -43,41 +44,23 @@ return {
     end,
   },
 
-  -- blink.cmp integration
-  --
-  -- FIXME: this currently doesn't work properly
-  -- {
-  --   "saghen/blink.cmp",
-  --   optional = true,
-  --   opts = {
-  --     sources = {
-  --       compat = vim.g.ai_cmp and { "supermaven" } or nil,
-  --     },
-  --   },
-  --   dependencies = {
-  --     "supermaven-nvim",
-  --     vim.g.ai_cmp and "saghen/blink.compat" or nil,
-  --   },
-  -- },
-  --
-  -- Disabble cmp integration for now
-  {
+  vim.g.ai_cmp and {
     "saghen/blink.cmp",
     optional = true,
-    ---@module 'blink.cmp'
-    ---@type blink.cmp.Config
+    dependencies = { "supermaven-nvim", "saghen/blink.compat" },
     opts = {
-      completion = { ghost_text = { enabled = false } },
-    },
-    dependencies = {
-      {
-        "supermaven-nvim",
-        opts = {
-          disable_inline_completion = false,
+      sources = {
+        compat = { "supermaven" },
+        providers = {
+          supermaven = {
+            kind = "Supermaven",
+            score_offset = 100,
+            async = true,
+          },
         },
       },
     },
-  },
+  } or nil,
 
   {
     "nvim-lualine/lualine.nvim",