+++ lua/lazyvim/plugins/extras/coding/blink.lua
@@ -1,3 +1,4 @@
+---@diagnostic disable: missing-fields
 if lazyvim_docs then
   -- set to `true` to follow the main branch
   -- you need to have a working rust toolchain to build the plugin
@@ -8,6 +9,7 @@ end
 return {
   {
     "hrsh7th/nvim-cmp",
+    optional = true,
     enabled = false,
   },
   {
@@ -17,6 +19,7 @@ return {
     opts_extend = {
       "sources.completion.enabled_providers",
       "sources.compat",
+      "sources.default",
     },
     dependencies = {
       "rafamadriz/friendly-snippets",
@@ -33,18 +36,31 @@ return {
     ---@module 'blink.cmp'
     ---@type blink.cmp.Config
     opts = {
-      highlight = {
+      snippets = {
+        expand = function(snippet, _)
+          return LazyVim.cmp.expand(snippet)
+        end,
+      },
+      appearance = {
         -- sets the fallback highlight groups to nvim-cmp's highlight groups
         -- useful for when your theme doesn't support blink.cmp
         -- will be removed in a future release, assuming themes add support
         use_nvim_cmp_as_default = false,
+        -- set to 'mono' for 'Nerd Font Mono' or 'normal' for 'Nerd Font'
+        -- adjusts spacing to ensure icons are aligned
+        nerd_font_variant = "mono",
       },
-      -- set to 'mono' for 'Nerd Font Mono' or 'normal' for 'Nerd Font'
-      -- adjusts spacing to ensure icons are aligned
-      nerd_font_variant = "mono",
       completion = {
+        accept = {
+          -- experimental auto-brackets support
+          auto_brackets = {
+            enabled = true,
+          },
+        },
         menu = {
-          winblend = vim.o.pumblend,
+          draw = {
+            treesitter = { "lsp" },
+          },
         },
         documentation = {
           auto_show = true,
@@ -55,33 +71,26 @@ return {
         },
       },
 
-      -- experimental auto-brackets support
-      accept = { auto_brackets = { enabled = true } },
-
       -- experimental signature help support
-      -- trigger = { signature_help = { enabled = true } }
+      -- signature = { enabled = true },
+
       sources = {
         -- adding any nvim-cmp sources here will enable them
         -- with blink.compat
         compat = {},
-        completion = {
-          -- remember to enable your providers here
-          enabled_providers = { "lsp", "path", "snippets", "buffer" },
-        },
+        default = { "lsp", "path", "snippets", "buffer" },
+        cmdline = {},
       },
 
       keymap = {
         preset = "enter",
-        ["<Tab>"] = {
-          LazyVim.cmp.map({ "snippet_forward", "ai_accept" }),
-          "fallback",
-        },
+        ["<C-y>"] = { "select_and_accept" },
       },
     },
     ---@param opts blink.cmp.Config | { sources: { compat: string[] } }
     config = function(_, opts)
       -- setup compat sources
-      local enabled = opts.sources.completion.enabled_providers
+      local enabled = opts.sources.default
       for _, source in ipairs(opts.sources.compat or {}) do
         opts.sources.providers[source] = vim.tbl_deep_extend(
           "force",
@@ -92,6 +101,65 @@ return {
           table.insert(enabled, source)
         end
       end
+
+      -- add ai_accept to <Tab> key
+      if not opts.keymap["<Tab>"] then
+        if opts.keymap.preset == "super-tab" then -- super-tab
+          opts.keymap["<Tab>"] = {
+            require("blink.cmp.keymap.presets")["super-tab"]["<Tab>"][1],
+            LazyVim.cmp.map({ "snippet_forward", "ai_accept" }),
+            "fallback",
+          }
+        else -- other presets
+          opts.keymap["<Tab>"] = {
+            LazyVim.cmp.map({ "snippet_forward", "ai_accept" }),
+            "fallback",
+          }
+        end
+      end
+
+      ---  NOTE: compat with latest version. Currenlty 0.7.6
+      if not vim.g.lazyvim_blink_main then
+        ---@diagnostic disable-next-line: inject-field
+        opts.sources.completion = opts.sources.completion or {}
+        opts.sources.completion.enabled_providers = enabled
+        if vim.tbl_get(opts, "completion", "menu", "draw", "treesitter") then
+          ---@diagnostic disable-next-line: assign-type-mismatch
+          opts.completion.menu.draw.treesitter = true
+        end
+      end
+
+      -- Unset custom prop to pass blink.cmp validation
+      opts.sources.compat = nil
+
+      -- check if we need to override symbol kinds
+      for _, provider in pairs(opts.sources.providers or {}) do
+        ---@cast provider blink.cmp.SourceProviderConfig|{kind?:string}
+        if provider.kind then
+          local CompletionItemKind = require("blink.cmp.types").CompletionItemKind
+          local kind_idx = #CompletionItemKind + 1
+
+          CompletionItemKind[kind_idx] = provider.kind
+          ---@diagnostic disable-next-line: no-unknown
+          CompletionItemKind[provider.kind] = kind_idx
+
+          ---@type fun(ctx: blink.cmp.Context, items: blink.cmp.CompletionItem[]): blink.cmp.CompletionItem[]
+          local transform_items = provider.transform_items
+          ---@param ctx blink.cmp.Context
+          ---@param items blink.cmp.CompletionItem[]
+          provider.transform_items = function(ctx, items)
+            items = transform_items and transform_items(ctx, items) or items
+            for _, item in ipairs(items) do
+              item.kind = kind_idx or item.kind
+            end
+            return items
+          end
+
+          -- Unset custom prop to pass blink.cmp validation
+          provider.kind = nil
+        end
+      end
+
       require("blink.cmp").setup(opts)
     end,
   },
@@ -101,7 +169,9 @@ return {
     "saghen/blink.cmp",
     opts = function(_, opts)
       opts.appearance = opts.appearance or {}
-      opts.appearance.kind_icons = LazyVim.config.icons.kinds
+      opts.appearance.kind_icons = vim.tbl_extend("keep", {
+        Color = "██", -- Use block instead of icon for color items to make swatches more usable
+      }, LazyVim.config.icons.kinds)
     end,
   },
 
@@ -110,18 +180,13 @@ return {
     "saghen/blink.cmp",
     opts = {
       sources = {
-        completion = {
-          -- add lazydev to your completion providers
-          enabled_providers = { "lazydev" },
-        },
+        -- add lazydev to your completion providers
+        default = { "lazydev" },
         providers = {
-          lsp = {
-            -- dont show LuaLS require statements when lazydev has items
-            fallback_for = { "lazydev" },
-          },
           lazydev = {
             name = "LazyDev",
             module = "lazydev.integrations.blink",
+            score_offset = 100, -- show at a higher priority than lsp
           },
         },
       },