+++ lua/lazyvim/plugins/extras/lang/sql.lua
@@ -130,9 +130,7 @@ return {
     optional = true,
     opts = {
       sources = {
-        completion = {
-          enabled_providers = { "dadbod" },
-        },
+        default = { "dadbod" },
         providers = {
           dadbod = { name = "Dadbod", module = "vim_dadbod_completion.blink" },
         },